fix(mvt): wgs84 coordinates for binary mvt
