+++ modules/mvt/test/index.ts
@@ -1,0 +1,3 @@
+import './mvt-loader.spec';
+import './lib/vector-tile-feature.spec';
+import './mvt-loader.spec';
