+++ modules/mvt/test/mvt-loader.spec.ts
@@ -1,11 +1,9 @@
-/** @typedef {import('@loaders.gl/schema').BinaryFeatures} BinaryFeatures */
+import type { BinaryFeatures } from '@loaders.gl/schema';
 import test from 'tape-promise/tape';
 import {MVTLoader} from '@loaders.gl/mvt';
 import {setLoaderOptions, fetchFile, parse, parseSync} from '@loaders.gl/core';
 import {geojsonToBinary, binaryToGeojson} from '@loaders.gl/gis';
 import {TEST_EXPORTS} from '../src/lib/binary-vector-tile/vector-tile-feature';
-
-const {classifyRings} = TEST_EXPORTS;
 
 const MVT_POINTS_DATA_URL = '@loaders.gl/mvt/test/data/points_4-2-6.mvt';
 const MVT_LINES_DATA_URL = '@loaders.gl/mvt/test/data/lines_2-2-1.mvt';
@@ -24,12 +22,6 @@ import decodedPolygonsGeometry from '@loaders.gl/mvt/test/results/decoded_mvt_po
 import decodedPointsGeoJSON from '@loaders.gl/mvt/test/results/decoded_mvt_points.json';
 import decodedLinesGeoJSON from '@loaders.gl/mvt/test/results/decoded_mvt_lines.json';
 import decodedPolygonsGeoJSON from '@loaders.gl/mvt/test/results/decoded_mvt_polygons.json';
-
-// Rings
-import ringsSingleRing from '@loaders.gl/mvt/test/data/rings_single_ring.json';
-import ringsRingAndHole from '@loaders.gl/mvt/test/data/rings_ring_and_hole.json';
-import ringsTwoRings from '@loaders.gl/mvt/test/data/rings_two_rings.json';
-import ringsZeroSizeHole from '@loaders.gl/mvt/test/data/rings_zero_size_hole.json';
 
 setLoaderOptions({
   _workerType: 'test'
@@ -228,6 +220,7 @@ const TEST_FILES = [
   MVT_POLYGON_ZERO_SIZE_HOLE_DATA_URL,
   MVT_MULTIPLE_LAYERS_DATA_URL
 ];
+
 for (const filename of TEST_FILES) {
   test(`geojson-to-binary generation is equivalent ${filename}`, async (t) => {
     const response = await fetchFile(filename);
@@ -300,38 +293,17 @@ test('Triangulation is supported', async (t) => {
   t.end();
 });
 
-test('Rings - single ring', async (t) => {
-  const geom = {...ringsSingleRing};
-  const classified = classifyRings(geom);
-  t.deepEqual(classified.areas, [[-0.02624368667602539]]);
-  t.deepEqual(classified.indices, [[0]]);
-  t.end();
-});
+test('Binary features with coordinates=wgs84', async (t) => {
+  const response = await fetchFile(WITH_FEATURE_ID);
+  const mvtArrayBuffer = await response.arrayBuffer();
 
-test('Rings - ring and hole', async (t) => {
-  const geom = {...ringsRingAndHole};
-  const classified = classifyRings(geom);
-  t.deepEqual(classified.areas, [[-0.02624368667602539, 0.001363515853881836]]);
-  t.deepEqual(classified.indices, [[0, 10]]);
-  t.end();
-});
+  const binary = await parse(mvtArrayBuffer, MVTLoader, {
+    mvt: {coordinates: 'wgs84'},
+    gis: {format: 'binary'}
+  }) as BinaryFeatures;
+  t.ok(binary.points?.positions.value.every(x => Number.isFinite(x)), 'positions are valid numbers');
+  t.ok(binary.lines?.positions.value.every(x => Number.isFinite(x)), 'positions are valid numbers');
+  t.ok(binary.polygons?.positions.value.every(x => Number.isFinite(x)), 'positions are valid numbers');
 
-test('Rings - two rings', async (t) => {
-  const geom = {...ringsTwoRings};
-  const classified = classifyRings(geom);
-  t.deepEqual(classified.areas, [[-0.02624368667602539], [-0.001363515853881836]]);
-  t.deepEqual(classified.indices, [[0], [10]]);
-  t.end();
-});
-
-test('Rings - zero sized hole', async (t) => {
-  // In addition to checking the result,
-  // verify that the data array is shortened
-  const geom = {...ringsZeroSizeHole};
-  t.equal(geom.data.length, 20);
-  const classified = classifyRings(geom);
-  t.deepEqual(classified.areas, [[-0.44582176208496094]]);
-  t.deepEqual(classified.indices, [[0]]);
-  t.equal(classified.data.length, 12);
   t.end();
 });
